+++ src/main/java/com/hyperlink/server/global/filter/AuthenticationFilter.java
@@ -45,7 +45,7 @@ public class AuthenticationFilter implements Filter {
       try {
         final String authorizationHeader = httpRequest.getHeader(HttpHeaders.AUTHORIZATION);
         final String accessToken = authTokenExtractor.extractToken(authorizationHeader);
-        jwtTokenProvider.validateExpiredToken(accessToken);
+        authTokenExtractor.validateExpiredToken(accessToken);
       } catch (TokenExpiredException e) {
         httpResponse.setStatus(HttpStatus.UNAUTHORIZED.value());
         return;
